Better split OmniKit automation hooks
