+++ features/omni-kit/automation/actions/common.ts
@@ -1,4 +1,4 @@
-import type { OmniAutomationSimulationResponse } from 'features/omni-kit/contexts'
+import type { OmniAutomationSimulationResponse } from 'features/omni-kit/types'
 
 export const defaultAutomationActionPromise = Promise.resolve<
   OmniAutomationSimulationResponse | undefined
