+++ features/omni-kit/automation/hooks/useOmniAutomationTxHandler.ts
@@ -8,10 +8,10 @@ import { ethers } from 'ethers'
 import { AutomationFeatures } from 'features/automation/common/types'
 import type { OmniGetAutomationDataParams } from 'features/omni-kit/automation/helpers'
 import { getOmniAutomationParameters } from 'features/omni-kit/automation/helpers'
-import type { OmniAutomationSimulationResponse } from 'features/omni-kit/contexts'
 import { useOmniGeneralContext, useOmniProductContext } from 'features/omni-kit/contexts'
 import type { OmniTxData } from 'features/omni-kit/hooks'
 import { estimateOmniGas$, sendOmniTransaction$ } from 'features/omni-kit/observables'
+import type { OmniAutomationSimulationResponse } from 'features/omni-kit/types'
 import { handleTransaction } from 'helpers/handleTransaction'
 import { useObservable } from 'helpers/observableHook'
 import { useDebouncedEffect } from 'helpers/useDebouncedEffect'
