+++ features/omni-kit/contexts/OmniProductContext.tsx
@@ -1,323 +1,22 @@
-import type { LendingPosition, Strategy, SupplyPosition } from '@oasisdex/dma-library'
-import type BigNumber from 'bignumber.js'
 import { useProductContext } from 'components/context/ProductContextProvider'
-import type { DetailsSectionNotificationItem } from 'components/DetailsSectionNotification'
-import type { ItemProps } from 'components/infoSection/Item'
-import type { SidebarSectionHeaderSelectItem } from 'components/sidebar/SidebarSectionHeaderSelect'
-import type { HeadlineDetailsProp } from 'components/vault/VaultHeadlineDetails'
-import { AutomationFeatures } from 'features/automation/common/types'
 import { useOmniGeneralContext } from 'features/omni-kit/contexts'
 import { getOmniValidations } from 'features/omni-kit/helpers'
+import { useOmniInitialization } from 'features/omni-kit/hooks'
 import { formatSwapData } from 'features/omni-kit/protocols/ajna/helpers'
-import {
-  getAutomationAutoBSFormDefaults,
-  useOmniAutomationAutoBSFormReducto,
-} from 'features/omni-kit/state/automation/auto-bs'
 import type {
-  OmniAutomationFormState,
-  useOmniAutomationFormReducto,
-} from 'features/omni-kit/state/automation/common'
-import {
-  getAutomationPartialTakeProfitFormDefaults,
-  useOmniAutomationPartialTakeProfitFormReducto,
-} from 'features/omni-kit/state/automation/partial-take-profit'
-import {
-  getAutomationStopLossFormDefaults,
-  useOmniStopLossAutomationFormReducto,
-} from 'features/omni-kit/state/automation/stop-loss'
-import {
-  getAutomationTrailingStopLossFormDefaults,
-  useOmniAutomationTrailingStopLossFormReducto,
-} from 'features/omni-kit/state/automation/trailing-stop-loss'
-import type { OmniBorrowFormState, useOmniBorrowFormReducto } from 'features/omni-kit/state/borrow'
-import type { OmniEarnFormState, useOmniEarnFormReducto } from 'features/omni-kit/state/earn'
-import type {
-  OmniMultiplyFormState,
-  useOmniMultiplyFormReducto,
-} from 'features/omni-kit/state/multiply'
-import type {
-  OmniGenericPosition,
+  OmniPositionSet,
   OmniSimulationCommon,
   OmniSimulationSwap,
-  OmniValidations,
+  ProductContextWithBorrow,
+  ProductContextWithEarn,
+  ProductContextWithMultiply,
+  ProductDetailsContextProviderProps,
 } from 'features/omni-kit/types'
 import { OmniProductType } from 'features/omni-kit/types'
-import type { PositionHistoryEvent } from 'features/positionHistory/types'
 import { useObservable } from 'helpers/observableHook'
-import type {
-  AutoBuyTriggersWithDecodedParams,
-  AutoSellTriggersWithDecodedParams,
-  GetTriggersResponse,
-  PartialTakeProfitTriggers,
-  PartialTakeProfitTriggersWithDecodedParams,
-  SetupBasicAutoResponse,
-  SetupBasicStopLossResponse,
-  SetupPartialTakeProfitResponse,
-  SetupTrailingStopLossResponse,
-  StopLossTriggersWithDecodedParams,
-  TrailingStopLossTriggersWithDecodedParams,
-} from 'helpers/triggers'
 import { useAccount } from 'helpers/useAccount'
-import type { Dispatch, FC, PropsWithChildren, ReactNode, SetStateAction } from 'react'
-import React, { useContext, useMemo, useState } from 'react'
-import type { Theme } from 'theme-ui'
-import type { CreatePositionEvent } from 'types/ethers-contracts/PositionCreated'
-
-interface OmniFeatureToggles {
-  safetySwitch: boolean
-  suppressValidation: boolean
-}
-
-interface OmniFilters {
-  flowStateFilter: (event: CreatePositionEvent) => boolean
-}
-
-interface CommonMetadata {
-  featureToggles: OmniFeatureToggles
-  filters: OmniFilters
-  notifications: DetailsSectionNotificationItem[]
-  validations: OmniValidations
-}
-
-export interface OmniLendingMetadataHandlers {}
-
-export interface OmniSupplyMetadataHandlers {
-  txSuccessEarnHandler?: () => void
-  customReset?: () => void
-}
-
-export interface AutomationMetadataValues {
-  flags: {
-    isStopLossEnabled: boolean
-    isTrailingStopLossEnabled: boolean
-    isAutoSellEnabled: boolean
-    isAutoBuyEnabled: boolean
-    isPartialTakeProfitEnabled: boolean
-  }
-  triggers: {
-    [AutomationFeatures.STOP_LOSS]?: StopLossTriggersWithDecodedParams
-    [AutomationFeatures.TRAILING_STOP_LOSS]?: TrailingStopLossTriggersWithDecodedParams
-    [AutomationFeatures.AUTO_SELL]?: AutoSellTriggersWithDecodedParams
-    [AutomationFeatures.AUTO_BUY]?: AutoBuyTriggersWithDecodedParams
-    [AutomationFeatures.PARTIAL_TAKE_PROFIT]?: PartialTakeProfitTriggersWithDecodedParams
-    [AutomationFeatures.CONSTANT_MULTIPLE]?: PartialTakeProfitTriggers
-    [AutomationFeatures.AUTO_TAKE_PROFIT]?: PartialTakeProfitTriggers
-  }
-  resolved: {
-    activeUiDropdown: AutomationFeatures
-    activeForm: ProductContextAutomationForm
-    isProtection: boolean
-    isOptimization: boolean
-    isFormEmpty: boolean
-  }
-  simulation: AutomationMetadataValuesSimulation
-}
-
-interface CommonMetadataValues {
-  footerColumns: number
-  headline?: string
-  headlineDetails?: HeadlineDetailsProp[]
-  isHeadlineDetailsLoading?: boolean
-  interestRate: BigNumber
-  isFormEmpty: boolean
-  sidebarTitle?: string
-  automation?: AutomationMetadataValues
-}
-interface CommonMetadataElements {
-  faq: ReactNode
-  overviewBanner?: ReactNode
-  positionBanner?: ReactNode
-  overviewContent: ReactNode
-  overviewFooter: ReactNode
-  overviewWithSimulation?: boolean
-  sidebarContent?: ReactNode
-  riskSidebar?: ReactNode
-}
-
-export type ShouldShowDynamicLtvMetadata = (params: { includeCache: boolean }) => boolean
-
-export type LendingMetadata = CommonMetadata & {
-  handlers?: OmniLendingMetadataHandlers
-  values: CommonMetadataValues & {
-    afterAvailableToBorrow: BigNumber | undefined
-    afterBuyingPower: BigNumber | undefined
-    afterPositionDebt: BigNumber | undefined
-    changeVariant: 'positive' | 'negative'
-    withdrawMax: BigNumber
-    debtMax: BigNumber
-    debtMin: BigNumber
-    paybackMax: BigNumber
-    shouldShowDynamicLtv: ShouldShowDynamicLtvMetadata
-    maxSliderAsMaxLtv?: boolean
-  }
-  elements: CommonMetadataElements & {
-    highlighterOrderInformation?: ReactNode
-  }
-  theme?: Theme
-}
-
-export type SupplyMetadata = CommonMetadata & {
-  handlers: OmniSupplyMetadataHandlers
-  values: CommonMetadataValues & {
-    earnAfterWithdrawMax?: BigNumber
-    earnWithdrawMax: BigNumber
-    extraDropdownItems?: SidebarSectionHeaderSelectItem[]
-    withAdjust?: boolean
-  }
-  elements: CommonMetadataElements & {
-    earnExtraUiDropdownContent?: ReactNode
-    earnFormOrder: ReactNode
-    earnFormOrderAsElement: FC
-    extraEarnInput?: ReactNode
-    extraEarnInputDeposit?: ReactNode
-    extraEarnInputWithdraw?: ReactNode
-  }
-  theme?: Theme
-}
-
-export type OmniMetadataParams =
-  | Omit<ProductContextWithBorrow, 'dynamicMetadata'>
-  | Omit<ProductContextWithEarn, 'dynamicMetadata'>
-  | Omit<ProductContextWithMultiply, 'dynamicMetadata'>
-
-export type GetOmniMetadata = (_: OmniMetadataParams) => LendingMetadata | SupplyMetadata
-
-interface WithAutomation {
-  automationFormReducto: typeof useOmniAutomationFormReducto
-  automationFormDefaults: Partial<OmniAutomationFormState>
-  positionTriggers: GetTriggersResponse
-}
-
-interface ProductContextProviderPropsWithBorrow extends WithAutomation {
-  formDefaults: Partial<OmniBorrowFormState>
-  formReducto: typeof useOmniBorrowFormReducto
-  getDynamicMetadata: GetOmniMetadata
-  position: LendingPosition
-  positionAuction: unknown
-  positionHistory: PositionHistoryEvent[]
-  productType: OmniProductType.Borrow
-}
-
-interface ProductContextProviderPropsWithEarn extends WithAutomation {
-  formDefaults: Partial<OmniEarnFormState>
-  formReducto: typeof useOmniEarnFormReducto
-  getDynamicMetadata: GetOmniMetadata
-  position: SupplyPosition
-  positionAuction: unknown
-  positionHistory: PositionHistoryEvent[]
-  productType: OmniProductType.Earn
-}
-
-interface ProductContextProviderPropsWithMultiply extends WithAutomation {
-  formDefaults: Partial<OmniMultiplyFormState>
-  formReducto: typeof useOmniMultiplyFormReducto
-  getDynamicMetadata: GetOmniMetadata
-  position: LendingPosition
-  positionAuction: unknown
-  positionHistory: PositionHistoryEvent[]
-  productType: OmniProductType.Multiply
-}
-
-type ProductDetailsContextProviderProps =
-  | ProductContextProviderPropsWithBorrow
-  | ProductContextProviderPropsWithEarn
-  | ProductContextProviderPropsWithMultiply
-
-interface PositionSet<Position> {
-  position: Position
-  simulation?: Position
-}
-
-interface OmniValidationMessage {
-  name: string
-  data?: { [key: string]: string | number }
-}
-
-export type OmniSimulationData<Position> = Strategy<Position>['simulation'] & {
-  errors: OmniValidationMessage[]
-  warnings: OmniValidationMessage[]
-  notices: OmniValidationMessage[]
-  successes: OmniValidationMessage[]
-}
-
-interface ProductContextPosition<Position, Auction> {
-  cachedPosition?: PositionSet<Position>
-  currentPosition: PositionSet<Position>
-  swap?: {
-    current?: OmniSimulationSwap
-    cached?: OmniSimulationSwap
-  }
-  isSimulationLoading?: boolean
-  resolvedId?: string
-  setCachedPosition: (positionSet: PositionSet<OmniGenericPosition>) => void
-  setIsLoadingSimulation: Dispatch<SetStateAction<boolean>>
-  setSimulation: Dispatch<SetStateAction<OmniSimulationData<OmniGenericPosition> | undefined>>
-  setCachedSwap: (swap: OmniSimulationSwap) => void
-  positionAuction: Auction
-  history: PositionHistoryEvent[]
-  simulationCommon: OmniSimulationCommon
-}
-
-export type OmniAutomationSimulationResponse =
-  | SetupBasicAutoResponse
-  | SetupBasicStopLossResponse
-  | SetupTrailingStopLossResponse
-  | SetupPartialTakeProfitResponse
-
-type AutomationMetadataValuesSimulation = OmniAutomationSimulationResponse['simulation']
-
-export interface ProductContextAutomationForms {
-  stopLoss: ReturnType<typeof useOmniStopLossAutomationFormReducto>
-  trailingStopLoss: ReturnType<typeof useOmniAutomationTrailingStopLossFormReducto>
-  autoSell: ReturnType<typeof useOmniAutomationAutoBSFormReducto>
-  autoBuy: ReturnType<typeof useOmniAutomationAutoBSFormReducto>
-  partialTakeProfit: ReturnType<typeof useOmniAutomationPartialTakeProfitFormReducto>
-  autoTakeProfit: ReturnType<typeof useOmniAutomationPartialTakeProfitFormReducto>
-  constantMultiple: ReturnType<typeof useOmniAutomationPartialTakeProfitFormReducto>
-}
-
-export type ProductContextAutomationForm =
-  ProductContextAutomationForms[keyof ProductContextAutomationForms]
-
-interface ProductContextAutomation {
-  positionTriggers: GetTriggersResponse
-  automationForms: ProductContextAutomationForms
-  commonForm: ReturnType<typeof useOmniAutomationFormReducto>
-  simulationData?: OmniAutomationSimulationResponse
-  isSimulationLoading?: boolean
-  setIsLoadingSimulation: Dispatch<SetStateAction<boolean>>
-  setSimulation: Dispatch<SetStateAction<OmniAutomationSimulationResponse | undefined>>
-  cachedOrderInfoItems?: ItemProps[]
-  setCachedOrderInfoItems: Dispatch<SetStateAction<ItemProps[] | undefined>>
-}
-
-interface GenericProductContext<Position, Form, Auction, Metadata> {
-  form: Form
-  position: ProductContextPosition<Position, Auction>
-  dynamicMetadata: Metadata
-  automation: ProductContextAutomation
-}
-
-export type ProductContextWithBorrow = GenericProductContext<
-  LendingPosition,
-  ReturnType<typeof useOmniBorrowFormReducto>,
-  unknown,
-  LendingMetadata
->
-
-export type ProductContextWithEarn = GenericProductContext<
-  SupplyPosition,
-  ReturnType<typeof useOmniEarnFormReducto>,
-  unknown,
-  SupplyMetadata
->
-
-export type ProductContextWithMultiply = GenericProductContext<
-  LendingPosition,
-  ReturnType<typeof useOmniMultiplyFormReducto>,
-  unknown,
-  LendingMetadata
->
+import type { PropsWithChildren } from 'react'
+import React, { useContext, useMemo } from 'react'
 
 const borrowContext = React.createContext<ProductContextWithBorrow | undefined>(undefined)
 const earnContext = React.createContext<ProductContextWithEarn | undefined>(undefined)
@@ -390,31 +89,25 @@ export function OmniProductContextProvider({
   const form = formReducto(formDefaults)
 
   const automationForm = automationFormReducto(automationFormDefaults)
-  const stopLossForm = useOmniStopLossAutomationFormReducto(
-    getAutomationStopLossFormDefaults(positionTriggers),
-  )
-  const trailingStopLossForm = useOmniAutomationTrailingStopLossFormReducto(
-    getAutomationTrailingStopLossFormDefaults(positionTriggers),
-  )
-  const autoSellForm = useOmniAutomationAutoBSFormReducto(
-    getAutomationAutoBSFormDefaults(positionTriggers, AutomationFeatures.AUTO_SELL),
-  )
-  const autoBuyForm = useOmniAutomationAutoBSFormReducto(
-    getAutomationAutoBSFormDefaults(positionTriggers, AutomationFeatures.AUTO_BUY),
-  )
-  const partialTakeProfitForm = useOmniAutomationPartialTakeProfitFormReducto(
-    getAutomationPartialTakeProfitFormDefaults(positionTriggers),
-  )
-
-  const automationForms = {
-    stopLoss: stopLossForm,
-    trailingStopLoss: trailingStopLossForm,
-    autoSell: autoSellForm,
-    autoBuy: autoBuyForm,
-    partialTakeProfit: partialTakeProfitForm,
-    constantMultiple: partialTakeProfitForm,
-    autoTakeProfit: partialTakeProfitForm,
-  }
+  const {
+    automationForms,
+    cachedPosition,
+    setCachedPosition,
+    cachedSwap,
+    setCachedSwap,
+    simulation,
+    setSimulation,
+    isSimulationLoading,
+    setIsLoadingSimulation,
+    automationSimulationData,
+    setAutomationSimulationData,
+    isAutomationSimulationLoading,
+    setAutomationIsLoadingSimulation,
+    cachedAutomationOrderInfoItems,
+    setCachedAutomationOrderInfoItems,
+  } = useOmniInitialization<typeof position>({
+    positionTriggers,
+  })
 
   const { state } = form
   const { state: automationState } = automationForm
@@ -422,19 +115,6 @@ export function OmniProductContextProvider({
   const [positionIdFromDpmProxyData] = useObservable(
     useMemo(() => positionIdFromDpmProxy$(state.dpmAddress), [state.dpmAddress]),
   )
-
-  // TODO these could be potentially generalized within single hook
-  const [cachedPosition, setCachedPosition] = useState<PositionSet<typeof position>>()
-  const [cachedSwap, setCachedSwap] = useState<OmniSimulationSwap>()
-  const [simulation, setSimulation] = useState<OmniSimulationData<typeof position>>()
-  const [isSimulationLoading, setIsLoadingSimulation] = useState(false)
-  // TODO these could be potentially generalized within single hook
-
-  const [automationSimulationData, setAutomationSimulationData] =
-    useState<OmniAutomationSimulationResponse>()
-  const [isAutomationSimulationLoading, setAutomationIsLoadingSimulation] = useState(false)
-  const [cachedAutomationOrderInfoItems, setCachedAutomationOrderInfoItems] =
-    useState<ItemProps[]>()
 
   // We need to determine the direction of the swap based on change in position risk
   let isIncreasingPositionRisk = true
@@ -474,7 +154,7 @@ export function OmniProductContextProvider({
             state,
           }),
         },
-        setCachedPosition: (positionSet: PositionSet<typeof position>) =>
+        setCachedPosition: (positionSet: OmniPositionSet<typeof position>) =>
           setCachedPosition(positionSet),
         setIsLoadingSimulation,
         setSimulation,
