+++ features/omni-kit/helpers/omniMetadataSupplyHandlerGuard.ts
@@ -1,7 +1,7 @@
 import type {
   OmniLendingMetadataHandlers,
   OmniSupplyMetadataHandlers,
-} from 'features/omni-kit/contexts'
+} from 'features/omni-kit/types'
 
 export const omniMetadataSupplyHandlerGuard = (
   handlers: OmniLendingMetadataHandlers | OmniSupplyMetadataHandlers | undefined,
