+++ features/omni-kit/hooks/index.ts
@@ -7,3 +7,4 @@ export * from './useOmniTransition'
 export * from './useOmniTxHandler'
 export * from './useOmniYieldLoopHeadline'
 export * from './useOmniAutomationSidebarTitle'
+export * from './useOmniInitialization'
