+++ features/omni-kit/types.ts
@@ -1,18 +1,53 @@
-import type { LendingPosition, SupplyPosition, Swap, SwapData } from '@oasisdex/dma-library'
+import type {
+  LendingPosition,
+  Strategy,
+  SupplyPosition,
+  Swap,
+  SwapData,
+} from '@oasisdex/dma-library'
 import type BigNumber from 'bignumber.js'
 import { NetworkIds } from 'blockchain/networks'
 import type { Tickers } from 'blockchain/prices.types'
 import type { DetailsSectionNotificationItem } from 'components/DetailsSectionNotification'
+import type { ItemProps } from 'components/infoSection/Item'
+import type { SidebarSectionHeaderSelectItem } from 'components/sidebar/SidebarSectionHeaderSelect'
+import type { HeadlineDetailsProp } from 'components/vault/VaultHeadlineDetails'
 import type { AutomationFeatures } from 'features/automation/common/types'
 import type { DpmPositionData } from 'features/omni-kit/observables'
-import type { OmniBorrowFormState } from 'features/omni-kit/state/borrow'
-import type { OmniEarnFormState } from 'features/omni-kit/state/earn'
-import type { OmniMultiplyFormState } from 'features/omni-kit/state/multiply'
-import type { GetTriggersResponse } from 'helpers/triggers'
+import type { useOmniAutomationAutoBSFormReducto } from 'features/omni-kit/state/automation/auto-bs'
+import type {
+  OmniAutomationFormState,
+  useOmniAutomationFormReducto,
+} from 'features/omni-kit/state/automation/common'
+import type { useOmniAutomationPartialTakeProfitFormReducto } from 'features/omni-kit/state/automation/partial-take-profit'
+import type { useOmniStopLossAutomationFormReducto } from 'features/omni-kit/state/automation/stop-loss'
+import type { useOmniAutomationTrailingStopLossFormReducto } from 'features/omni-kit/state/automation/trailing-stop-loss'
+import type { OmniBorrowFormState, useOmniBorrowFormReducto } from 'features/omni-kit/state/borrow'
+import type { OmniEarnFormState, useOmniEarnFormReducto } from 'features/omni-kit/state/earn'
+import type {
+  OmniMultiplyFormState,
+  useOmniMultiplyFormReducto,
+} from 'features/omni-kit/state/multiply'
+import type { PositionHistoryEvent } from 'features/positionHistory/types'
+import type {
+  AutoBuyTriggersWithDecodedParams,
+  AutoSellTriggersWithDecodedParams,
+  GetTriggersResponse,
+  PartialTakeProfitTriggers,
+  PartialTakeProfitTriggersWithDecodedParams,
+  SetupBasicAutoResponse,
+  SetupBasicStopLossResponse,
+  SetupPartialTakeProfitResponse,
+  SetupTrailingStopLossResponse,
+  StopLossTriggersWithDecodedParams,
+  TrailingStopLossTriggersWithDecodedParams,
+} from 'helpers/triggers'
 import type { TxError } from 'helpers/types'
 import type { LendingProtocolLabel } from 'lendingProtocols'
 import { LendingProtocol } from 'lendingProtocols'
-import type { CreatePositionEvent } from 'types/ethers-contracts/AjnaProxyActions'
+import type { Dispatch, FC, ReactNode, SetStateAction } from 'react'
+import type { Theme } from 'theme-ui'
+import type { CreatePositionEvent } from 'types/ethers-contracts/PositionCreated'
 
 const omniSupportedNetworkIds = [
   NetworkIds.ARBITRUMMAINNET,
@@ -292,3 +327,259 @@ export interface OmniExtraTokenData {
     price: BigNumber
   }
 }
+
+interface OmniFeatureToggles {
+  safetySwitch: boolean
+  suppressValidation: boolean
+}
+
+interface OmniFilters {
+  flowStateFilter: (event: CreatePositionEvent) => boolean
+}
+
+interface CommonMetadata {
+  featureToggles: OmniFeatureToggles
+  filters: OmniFilters
+  notifications: DetailsSectionNotificationItem[]
+  validations: OmniValidations
+}
+
+export interface OmniLendingMetadataHandlers {}
+
+export interface OmniSupplyMetadataHandlers {
+  txSuccessEarnHandler?: () => void
+  customReset?: () => void
+}
+
+export interface AutomationMetadataValues {
+  flags: {
+    isStopLossEnabled: boolean
+    isTrailingStopLossEnabled: boolean
+    isAutoSellEnabled: boolean
+    isAutoBuyEnabled: boolean
+    isPartialTakeProfitEnabled: boolean
+  }
+  triggers: {
+    [AutomationFeatures.STOP_LOSS]?: StopLossTriggersWithDecodedParams
+    [AutomationFeatures.TRAILING_STOP_LOSS]?: TrailingStopLossTriggersWithDecodedParams
+    [AutomationFeatures.AUTO_SELL]?: AutoSellTriggersWithDecodedParams
+    [AutomationFeatures.AUTO_BUY]?: AutoBuyTriggersWithDecodedParams
+    [AutomationFeatures.PARTIAL_TAKE_PROFIT]?: PartialTakeProfitTriggersWithDecodedParams
+    [AutomationFeatures.CONSTANT_MULTIPLE]?: PartialTakeProfitTriggers
+    [AutomationFeatures.AUTO_TAKE_PROFIT]?: PartialTakeProfitTriggers
+  }
+  resolved: {
+    activeUiDropdown: AutomationFeatures
+    activeForm: ProductContextAutomationForm
+    isProtection: boolean
+    isOptimization: boolean
+    isFormEmpty: boolean
+  }
+  simulation: AutomationMetadataValuesSimulation
+}
+
+interface CommonMetadataValues {
+  footerColumns: number
+  headline?: string
+  headlineDetails?: HeadlineDetailsProp[]
+  isHeadlineDetailsLoading?: boolean
+  interestRate: BigNumber
+  isFormEmpty: boolean
+  sidebarTitle?: string
+  automation?: AutomationMetadataValues
+}
+interface CommonMetadataElements {
+  faq: ReactNode
+  overviewBanner?: ReactNode
+  positionBanner?: ReactNode
+  overviewContent: ReactNode
+  overviewFooter: ReactNode
+  overviewWithSimulation?: boolean
+  sidebarContent?: ReactNode
+  riskSidebar?: ReactNode
+}
+
+export type ShouldShowDynamicLtvMetadata = (params: { includeCache: boolean }) => boolean
+
+export type LendingMetadata = CommonMetadata & {
+  handlers?: OmniLendingMetadataHandlers
+  values: CommonMetadataValues & {
+    afterAvailableToBorrow: BigNumber | undefined
+    afterBuyingPower: BigNumber | undefined
+    afterPositionDebt: BigNumber | undefined
+    changeVariant: 'positive' | 'negative'
+    withdrawMax: BigNumber
+    debtMax: BigNumber
+    debtMin: BigNumber
+    paybackMax: BigNumber
+    shouldShowDynamicLtv: ShouldShowDynamicLtvMetadata
+    maxSliderAsMaxLtv?: boolean
+  }
+  elements: CommonMetadataElements & {
+    highlighterOrderInformation?: ReactNode
+  }
+  theme?: Theme
+}
+
+export type SupplyMetadata = CommonMetadata & {
+  handlers: OmniSupplyMetadataHandlers
+  values: CommonMetadataValues & {
+    earnAfterWithdrawMax?: BigNumber
+    earnWithdrawMax: BigNumber
+    extraDropdownItems?: SidebarSectionHeaderSelectItem[]
+    withAdjust?: boolean
+  }
+  elements: CommonMetadataElements & {
+    earnExtraUiDropdownContent?: ReactNode
+    earnFormOrder: ReactNode
+    earnFormOrderAsElement: FC
+    extraEarnInput?: ReactNode
+    extraEarnInputDeposit?: ReactNode
+    extraEarnInputWithdraw?: ReactNode
+  }
+  theme?: Theme
+}
+
+export type OmniMetadataParams =
+  | Omit<ProductContextWithBorrow, 'dynamicMetadata'>
+  | Omit<ProductContextWithEarn, 'dynamicMetadata'>
+  | Omit<ProductContextWithMultiply, 'dynamicMetadata'>
+
+export type GetOmniMetadata = (_: OmniMetadataParams) => LendingMetadata | SupplyMetadata
+
+export interface WithAutomation {
+  automationFormReducto: typeof useOmniAutomationFormReducto
+  automationFormDefaults: Partial<OmniAutomationFormState>
+  positionTriggers: GetTriggersResponse
+}
+
+export interface ProductContextProviderPropsWithBorrow extends WithAutomation {
+  formDefaults: Partial<OmniBorrowFormState>
+  formReducto: typeof useOmniBorrowFormReducto
+  getDynamicMetadata: GetOmniMetadata
+  position: LendingPosition
+  positionAuction: unknown
+  positionHistory: PositionHistoryEvent[]
+  productType: OmniProductType.Borrow
+}
+
+export interface ProductContextProviderPropsWithEarn extends WithAutomation {
+  formDefaults: Partial<OmniEarnFormState>
+  formReducto: typeof useOmniEarnFormReducto
+  getDynamicMetadata: GetOmniMetadata
+  position: SupplyPosition
+  positionAuction: unknown
+  positionHistory: PositionHistoryEvent[]
+  productType: OmniProductType.Earn
+}
+
+export interface ProductContextProviderPropsWithMultiply extends WithAutomation {
+  formDefaults: Partial<OmniMultiplyFormState>
+  formReducto: typeof useOmniMultiplyFormReducto
+  getDynamicMetadata: GetOmniMetadata
+  position: LendingPosition
+  positionAuction: unknown
+  positionHistory: PositionHistoryEvent[]
+  productType: OmniProductType.Multiply
+}
+
+export type ProductDetailsContextProviderProps =
+  | ProductContextProviderPropsWithBorrow
+  | ProductContextProviderPropsWithEarn
+  | ProductContextProviderPropsWithMultiply
+
+export interface OmniPositionSet<Position> {
+  position: Position
+  simulation?: Position
+}
+
+export interface OmniValidationMessage {
+  name: string
+  data?: { [key: string]: string | number }
+}
+
+export type OmniSimulationData<Position> = Strategy<Position>['simulation'] & {
+  errors: OmniValidationMessage[]
+  warnings: OmniValidationMessage[]
+  notices: OmniValidationMessage[]
+  successes: OmniValidationMessage[]
+}
+
+interface ProductContextPosition<Position, Auction> {
+  cachedPosition?: OmniPositionSet<Position>
+  currentPosition: OmniPositionSet<Position>
+  swap?: {
+    current?: OmniSimulationSwap
+    cached?: OmniSimulationSwap
+  }
+  isSimulationLoading?: boolean
+  resolvedId?: string
+  setCachedPosition: (positionSet: OmniPositionSet<OmniGenericPosition>) => void
+  setIsLoadingSimulation: Dispatch<SetStateAction<boolean>>
+  setSimulation: Dispatch<SetStateAction<OmniSimulationData<OmniGenericPosition> | undefined>>
+  setCachedSwap: (swap: OmniSimulationSwap) => void
+  positionAuction: Auction
+  history: PositionHistoryEvent[]
+  simulationCommon: OmniSimulationCommon
+}
+
+export type OmniAutomationSimulationResponse =
+  | SetupBasicAutoResponse
+  | SetupBasicStopLossResponse
+  | SetupTrailingStopLossResponse
+  | SetupPartialTakeProfitResponse
+
+export type AutomationMetadataValuesSimulation = OmniAutomationSimulationResponse['simulation']
+
+export interface ProductContextAutomationForms {
+  stopLoss: ReturnType<typeof useOmniStopLossAutomationFormReducto>
+  trailingStopLoss: ReturnType<typeof useOmniAutomationTrailingStopLossFormReducto>
+  autoSell: ReturnType<typeof useOmniAutomationAutoBSFormReducto>
+  autoBuy: ReturnType<typeof useOmniAutomationAutoBSFormReducto>
+  partialTakeProfit: ReturnType<typeof useOmniAutomationPartialTakeProfitFormReducto>
+  autoTakeProfit: ReturnType<typeof useOmniAutomationPartialTakeProfitFormReducto>
+  constantMultiple: ReturnType<typeof useOmniAutomationPartialTakeProfitFormReducto>
+}
+
+export type ProductContextAutomationForm =
+  ProductContextAutomationForms[keyof ProductContextAutomationForms]
+
+export interface ProductContextAutomation {
+  positionTriggers: GetTriggersResponse
+  automationForms: ProductContextAutomationForms
+  commonForm: ReturnType<typeof useOmniAutomationFormReducto>
+  simulationData?: OmniAutomationSimulationResponse
+  isSimulationLoading?: boolean
+  setIsLoadingSimulation: Dispatch<SetStateAction<boolean>>
+  setSimulation: Dispatch<SetStateAction<OmniAutomationSimulationResponse | undefined>>
+  cachedOrderInfoItems?: ItemProps[]
+  setCachedOrderInfoItems: Dispatch<SetStateAction<ItemProps[] | undefined>>
+}
+
+export interface GenericProductContext<Position, Form, Auction, Metadata> {
+  form: Form
+  position: ProductContextPosition<Position, Auction>
+  dynamicMetadata: Metadata
+  automation: ProductContextAutomation
+}
+
+export type ProductContextWithBorrow = GenericProductContext<
+  LendingPosition,
+  ReturnType<typeof useOmniBorrowFormReducto>,
+  unknown,
+  LendingMetadata
+>
+
+export type ProductContextWithEarn = GenericProductContext<
+  SupplyPosition,
+  ReturnType<typeof useOmniEarnFormReducto>,
+  unknown,
+  SupplyMetadata
+>
+
+export type ProductContextWithMultiply = GenericProductContext<
+  LendingPosition,
+  ReturnType<typeof useOmniMultiplyFormReducto>,
+  unknown,
+  LendingMetadata
+>
